requirements.txt
connexion == 2.2.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
Flask-PyMongo-2.3.0